+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add ocaml-version bound
 - Support OCaml 4.00
 - Bisect is not necessary to run the test suite
+- Add `Record.declare{1,2,3,4}` to build fixed-size layouts
+- Add `Record.layout_type` to use layout as types
 
 ## v0.1.0